requirements_dev.txt
aiobotocore==2.22.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioitertools==0.12.0
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
boto3==1.37.3
botocore==1.37.3
et_xmlfile==2.0.0
frozenlist==1.5.0
fsspec==2024.10.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
multidict==6.1.0
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
propcache==0.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
s3fs==2024.10.0
s3transfer==0.11.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wrapt==1.17.2
yarl==1.15.2
mlflow
dagshub